Fix crashing issues when plugin loading fails.  Also moves both EPICS plugins into a subdirectory to make the import situation a little less confusing.
