+++ pydm/data_plugins/epics_plugin.py
@@ -5,16 +5,16 @@
 import os
 EPICS_LIB = os.getenv("PYDM_EPICS_LIB")
 if EPICS_LIB == "pyepics":
-  from .pyepics_plugin import PyEPICSPlugin
+  from .epics_plugins.pyepics_plugin import PyEPICSPlugin
   EPICSPlugin = PyEPICSPlugin
 elif EPICS_LIB == "pyca":
-  from .psp_plugin import PSPPlugin
+  from .epics_plugins.psp_plugin import PSPPlugin
   EPICSPlugin = PSPPlugin
 else:
   try:
-    from .psp_plugin import PSPPlugin
+    from .epics_plugins.psp_plugin import PSPPlugin
     EPICSPlugin = PSPPlugin
   except ImportError:
-    from .pyepics_plugin import PyEPICSPlugin
+    from .epics_plugins.pyepics_plugin import PyEPICSPlugin
     EPICSPlugin = PyEPICSPlugin
 EPICSPlugin.protocol = "ca"
